requirements.txt
django
pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-rosetta
python-dotenv
channels[daphne]
channels-redis
django-filter